examples/research_flow.py
from datetime import datetime
from typing import Any, Dict, List

import httpx

from pynions import Flow, State


class ResearchState(State):
    query: str
    urls: List[str] = []
    content: Dict[str, str] = {}
    summary: str = ""


class ResearchFlow(Flow[ResearchState]):
    """Flow for researching a topic"""

    async def run(self):
        # Search for sources
        async with self.step("search"):
            self.state.urls = await self._search()

        # Scrape content
        async with self.step("scrape"):
            self.state.content = await self._scrape()

        # Generate summary
        async with self.step("summarize"):
            self.state.summary = await self._summarize()

        return {"summary": self.state.summary, "sources": self.state.urls}

    async def _search(self) -> List[str]:
        """Search for relevant URLs"""
        async with httpx.AsyncClient() as client:
            # Implement your search logic
            return ["http://example.com"]

    async def _scrape(self) -> Dict[str, str]:
        """Scrape content from URLs"""
        content = {}
        # Implement your scraping logic
        return content

    async def _summarize(self) -> str:
        """Summarize findings"""
        # Implement your summarization logic
        return "Summary of findings"
