pyproject.toml
# pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pynions"
version = "0.1.0"
description = "Simple AI automation framework"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
dependencies = [
    "litellm>=1.12.1",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0"
]